Modified the lib structure to remove ADM prefix from classes that can be used more generally.
Added extent/spread panning for loudspeaker and HOA playback.
